Add a kotti.depot_mountpoint setting.
