+++ CHANGES.txt
@@ -4,10 +4,15 @@ Change History
 1.3.0 - unreleased
 ------------------
 
+- Completely revised ``Depot`` integration.  See
+  http://kotti.readthedocs.org/en/latest/developing/advanced/blobs.html
+  for details.
+
 - Make ``kotti.resources.SaveDataMixin`` more versatile in that it now supports
   a ``data_filters`` attribute (or even a completely overridden ``data``
   attribute) on subclasses.  For an example for what this is useful, see the
-  new ``kotti_image`` package's readme and the Depot documentation (http://depot.readthedocs.org/en/latest/database.html#custom-behaviour-in-attachments).
+  new ``kotti_image`` package's readme and the Depot documentation
+  (http://depot.readthedocs.org/en/latest/database.html#custom-behaviour-in-attachments).
 
   ** These changes require a database migration. **
